+++ packages/client/src/initLowdefyContext.js
@@ -17,10 +17,10 @@
 import { urlQuery } from '@lowdefy/helpers';
 
 import callRequest from './callRequest.js';
-import setupLink from './setupLink.js';
-import createLinkComponent from './createLinkComponent.js';
 import createIcon from './createIcon.js';
 import createAuthMethods from './auth/createAuthMethods.js';
+import createLinkComponent from './createLinkComponent.js';
+import setupLink from './setupLink.js';
 
 const lowdefy = {
   _internal: {
@@ -32,6 +32,12 @@ const lowdefy = {
       return () => undefined;
     },
     link: () => undefined,
+    progress: {
+      state: {
+        progress: 0,
+      },
+      dispatch: () => undefined,
+    },
   },
   contexts: {},
   inputs: {},